项目更新完结，去除编译文件

+++ szk-club-subject/szk-club-infra/src/main/java/com/szk/subject/infra/basic/service/SubjectMappingService.java
@@ -8,18 +8,17 @@ import java.util.List;
  * 题目分类关系表(SubjectMapping)表服务接口
  *
  * @author makejava
- * @since 2025-02-16 23:27:44
+ * @since 2023-10-03 22:17:07
  */
 public interface SubjectMappingService {
 
     /**
      * 通过ID查询单条数据
      *
-     * @param id 主键
      * @return 实例对象
      */
-    SubjectMapping queryById(Long id);
-    
+    SubjectMapping queryById(int id);
+
     /**
      * 新增数据
      *
@@ -39,14 +38,12 @@ public interface SubjectMappingService {
     /**
      * 通过主键删除数据
      *
-     * @param id 主键
      * @return 是否成功
      */
-    boolean deleteById(Long id);
+    boolean deleteById(int id);
 
     /**
-     * 查询主键id
-     * @return 是否成功
+     * 查询标签id
      */
     List<SubjectMapping> queryLabelId(SubjectMapping subjectMapping);
 
@@ -54,4 +51,5 @@ public interface SubjectMappingService {
      * 批量插入
      */
     void batchInsert(List<SubjectMapping> mappingList);
+
 }
